requirements.txt
aws-cdk-lib==2.66.1
constructs>=10.0.0,<11.0.0
boto3>=1.24.56
botocore>=1.27.56
python-benedict==0.25.1
PyYAML==6.0
pre-commit
requests>=2.28.2